+++ CHANGELOG.md
@@ -15,6 +15,10 @@ version directory, and  then changes are introduced.
 
 - Add back registry domain configuration.
 
+### Changed
+
+- Change `kube-apiserver` image to include certs.
+
 ## [6.3.0] - 2020-06-22
 
 ## Changed